create a run object, rename run.py to process_run.py and process_runs.py to queue_runs.py
